# Gen/OpenApi.yaml
openapi: 3.0.0
info:
  description: System to keep track of customer orders
  version: 1.0.0
  title: Customer Orders API
  contact:
    email: SwingGuy1024@yahoo.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://localhost:8080
paths:
...
  /menuItem/{id}:
    get:
      summary: Gets a menuItem by ID
      description: Gets a MenuItem by its ID.
      operationId: getMenuItem
      parameters:
        - in: path
          name: id
          description: ID of menuItem to find
          required: true
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemDto'
        404:
          description: NotFound
...
security:
  - bearerAuth: [ ]
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    MenuItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        itemPrice:
          type: number
          description: Price. (BigDecimal)
        allowedOptions:
          type: array
          items:
            $ref: '#/components/schemas/MenuItemOptionDto'
      required:
        - name
        - itemPrice
    MenuItemOptionDto:
      type: object
      description: Option for a MenuItem
      properties:
        name:
          type: string
        deltaPrice:
          type: number
          description: Floating point price. Strings are easier to work with.
        id:
          type: integer
          format: int32
      required:
        - name
        - deltaPrice
    # CustomerOrderDto:
    #   properties:
    #     menuItem:
    #       $ref: '#/components/schemas/MenuItemDto'
    #     options:
    #       type: array
    #       items:
    #         $ref: '#/components/schemas/MenuItemDto'
    #     customer:
    #       $ref: '#/components/schemas/UserDto'
    #     status:
    #       type: string
    #       description: Order complete
    #       enum: [ PENDING, COMPLETE, CANCELLED ]
    #     id:
    #       type: integer
    #       format: int32
    #     orderTime:
    #       type: string
    #       format: date-time
    #     completeTime:
    #       type: string
    #       format: date-time
...
    UserDto:
      type: object
      properties:
#        id:
#          type: integer
#          format: int32
        username:
          type: string
          description: username
        password:
          type: string
          format: password
          description: password
        email:
          type: string
          description: email
        mobilePhone:
          type: string
          description: mobile phone number
        landPhone:
          type: string
          description: Land line phone number
      required:
        - username
        - password
...